# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation by Kang Fakhrur",
        "title": "HRIS Data Warehouse",
        "contact": {},
        "version": "V1.2412.081710"
    },
    "paths": {
...
        "/v1/dashboard/barchart-average-salary-per-department": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "JwtToken": []
                    }
                ],
                "description": "Get SalaryDistribution Bar Chart Average Salary Per Department",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SalaryDistribution"
                ],
                "summary": "Get SalaryDistribution Bar Chart Average Salary Per Department",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "department_id (int)",
                        "name": "department_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "emp_status_id (int)",
                        "name": "emp_status_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "position_id (int)",
                        "name": "position_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "manager_id (int)",
                        "name": "manager_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "gender (string)",
                        "name": "gender",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "state (string)",
                        "name": "state",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "start_date (format: 2006-01-02)",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "end_date (format: 2006-01-02)",
                        "name": "end_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/v1/dashboard/barchart-average-salary-per-position": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "JwtToken": []
                    }
                ],
                "description": "Get SalaryDistribution Bar Chart Average Salary Per Position with Employee Count",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SalaryDistribution"
                ],
                "summary": "Get SalaryDistribution Bar Chart Average Salary Per Position with Employee Count",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "department_id (int)",
                        "name": "department_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "emp_status_id (int)",
                        "name": "emp_status_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "position_id (int)",
                        "name": "position_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "manager_id (int)",
                        "name": "manager_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "gender (string)",
                        "name": "gender",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "state (string)",
                        "name": "state",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "start_date (format: 2006-01-02)",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "end_date (format: 2006-01-02)",
                        "name": "end_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/v1/dashboard/highest-lowest-salary": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "JwtToken": []
                    }
                ],
                "description": "Get Dashboard Highest and Lowest Salary",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SalaryDistribution"
                ],
                "summary": "Get Dashboard Highest and Lowest Salary",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "department_id (int)",
                        "name": "department_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "emp_status_id (int)",
                        "name": "emp_status_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "position_id (int)",
                        "name": "position_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "manager_id (int)",
                        "name": "manager_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "gender (string)",
                        "name": "gender",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "state (string)",
                        "name": "state",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "start_date (format: 2006-01-02)",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "end_date (format: 2006-01-02)",
                        "name": "end_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "JwtToken": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}